Safari-609.1.11.1/LayoutTests/fast/viewport/ios/shrink-to-fit-content-large-width-breakpoint-expected.txt

This test verifies that the shrink-to-fit-content heuristic doesn't induce more horizontal scrolling than we would otherwise have. To run the test manually, load the page and check that the bar almost entirely fits within the viewport, with no more than 480px of horizontal scrolling (on a 320px-wide device) and 20px of scrolling (on a 768px-wide device).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS minScale is 1
PASS 800 is >= document.scrollingElement.scrollWidth
PASS successfullyParsed is true

TEST COMPLETE
